Move all interface to Features.cs
